+++ day07/day07.cpp
@@ -65,7 +65,7 @@ Int Amplify2(std::vector<Int> const &code, std::array<Int, 5> const &sequence)
         Intcode{code}
     };
 
-    for (int i = 0; i != 5; ++i)
+    for (size_t i = 0; i != 5; ++i)
     {
         bool first = true;
 
@@ -93,7 +93,7 @@ Int Amplify2(std::vector<Int> const &code, std::array<Int, 5> const &sequence)
         for (auto &cpu : cpus)
         {
             auto v = cpu.RunUntilOuput(getParam);
-            
+
             if (v.has_value())
                 param = *v;
         }
